move nunit runtime provider to plugin
